some configs and responsive header
