+++ src/components/layout/Header/Header.css
@@ -1,98 +1,83 @@
 .header {
     background-color: #214ECF;
+    padding: 0 2%;
+    width: 100%;
     height: 145px;
-}
-
-.logo {
-    height: 70px;
-    width: auto;
-    margin-left: 50px;
-}
-.header-nav {
-    position: absolute;
-    top: 30px;
-    left: 50px;
-    right: 50px;
     display: flex;
     justify-content: space-between;
-    align-items: center;
-    z-index: 4;
-}
-
-.nav-left {
-    display: flex;
     align-items: center;
 }
 
 .nav-group {
     display: flex;
     align-items: center;
+    width: 65%;
 }
 
-
-/* Search icon */
 .nav-search {
     display: flex;
     align-items: center;
     justify-content: center;
-    width: 60px;
     height: 60px;
+    width: 60px;
     background: rgba(255, 255, 255, 0.8);
-    backdrop-filter: blur(10px);
-    border: 2px solid rgba(255, 255, 255, 0.5);
-    border-radius: 50px 50px 50px 50px;
-    color: #fff;
-    font-size: 1rem;
-    cursor: pointer;
+    border-radius: 999px;
 }
-
 .nav-search img {
-    width: 30px;
-    height: 30px;
-    object-fit: contain; 
+    width: 50%;
+    height: 50%; 
 }
 
-/* Nav links capsule */
+.logo {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin: auto;
+}
+.logo img {
+    height: 75px;
+}
+
 .nav-links {
     display: flex;
     align-items: center;
-    gap: 30px;
+    justify-content: center;
+    margin: auto;
     list-style: none;
-    height: 50px;
-    padding: 5px 30px;
     background: rgba(255, 255, 255, 0.8);
     backdrop-filter: blur(15px);
-    border: 2px solid rgba(255, 255, 255, 0.5);
-    margin: auto;
-    margin-left: 350px;
-    width: 450px;
-    border-radius: 50px 50px 50px 50px;
+    border-radius: 999px;
+    gap: 30px;
+    height: 60px;
+    padding: 5px 30px;
+    width: 480px;
     font-size: 20px;
 }
-
 
 .nav-links a {
     color: #000;
     text-decoration: none;
     font-weight: 500;
-    line-height: 1;
     margin: 9px;
 }
 
-/* Auth buttons */
+
+
 .nav-auth {
     display: flex;
-    gap: 10px;
+    align-items: center;
+    justify-content: right;
+    gap: 3%;
+    margin-right: 3.2%;
 }
-
 .nav-btn {
     background: #ffffff;
     color: #000;
     border: none;
-    padding: 8px 20px;
-    border-radius: 25px;
     cursor: pointer;
     font-weight: bold;
+    padding: 8px 20px;
+    border-radius: 25px;
     font-size: 35px;
     height: 85px;
     width: 190px;
