+++ src/components/layout/Header/Header.jsx
@@ -6,30 +6,29 @@ import SearchIcon from './searchicon.svg';
 const Header = () => {
     return (
         <header className="header">
-            <nav className="header-nav">
-                <div className="nav-left">
-                    <div className="nav-group">
-                        <div className="nav-search">
-                            <img src={SearchIcon} id="SearchIcon" />
-                        </div>
-                        <div className="logo">
-                            <img src="/assets/logo2.png" class="logo" alt="MealMaster" />
-                        </div>
-                        
-                        <ul className="nav-links">
-                            <li><a href="#">About</a></li>
-                            <li><a href="#">Pricing</a></li>
-                            <li><a href="#">Service</a></li>
-                            <li><a href="#">Community</a></li>
-                        </ul>
-                    </div>
+
+            <div className="nav-group">    
+                <div className="nav-search">
+                    <img src={SearchIcon} id="SearchIcon" />
                 </div>
 
-                <div className="nav-auth">
-                    <button className="nav-btn nav-signup">Sign up</button>
-                    <button className="nav-btn nav-login">Login</button>
+                <div className="logo">
+                    <img src="/assets/logo2.png" alt="MealMaster" />
                 </div>
-            </nav>
+                
+                <ul className="nav-links">
+                    <li><a href="#">About</a></li>
+                    <li><a href="#">Pricing</a></li>
+                    <li><a href="#">Service</a></li>
+                    <li><a href="#">Community</a></li>
+                </ul>
+            </div>
+
+            <div className="nav-auth">
+                <button className="nav-btn nav-signup">Sign up</button>
+                <button className="nav-btn nav-login">Login</button>
+            </div>
+            
         </header>
     );
 };
