+++ src/components/sections/HeroSection/HeroSection.css
@@ -3,22 +3,16 @@
 
 .hero-section {
     position: relative;
-    margin: 1.5%;
-    width: 97%;
     height: 60vh;
     overflow: hidden;
     color: #fff;
-    padding-bottom: 0;
-    padding-left: 0;
-    padding-right: 0;
-    background: linear-gradient(to bottom, #d6d6d6 0%, #ffffff 100%);
     border-radius: 30px 30px 30px 0px;
     font-family: 'Google Sans', sans-serif;
 }
 
 .hero-bg img {
     position: absolute;
-    width: 100vw;      /* thay vì 140% */
+    width: 140%;      /* thay vì 140% */
     height: 100%;
     object-fit: cover;
     z-index: 0;
@@ -37,7 +31,7 @@
     bottom: 0;
     left: 0;
     filter: brightness(0) invert(1); /* ép về trắng */
-    width: 105%;
+    width: 100%;
     height: 27%;
     z-index: 3;
     display: block;
