+++ src/components/sections/ProblemSection/ProblemSection.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./ProblemSection.css";
-import IconButton from "../../buttons/IconButton";
+import IconButton from "../../ui/IconButton/IconButton";
 
 import woman from "./woman.png";
 import question from "./question.png";
