+++ src/components/sections/index.js
@@ -2,12 +2,12 @@
 export { default as HeroSection } from "./HeroSection";
 export { default as ProblemSection } from "./ProblemSection";
 export { default as SolutionSection } from "./SolutionSection";
-
 export { default as DifferenceSection } from "./DifferenceSection";
 export { default as MarketStatsSection } from "./MarketStatsSection";
-export { default as StatementSection } from "./StatementSection";
 
+export { default as StatementSection } from "./StatementSection";
 export { default as FAQSection } from "./FAQSection";
+export { default as BannerSection } from "./BannerSection";
 export { default as JoinSection } from "./JoinSection";
 export { default as RecruitSection } from "./RecruitSection";
-
+export { default as RegisterBannerSection } from "./RegisterBannerSection";
