+++ src/components/ui/IconButton/index.js
@@ -1,0 +1,3 @@
+import IconButton from "./IconButton";
+
+export default IconButton;
