+++ src/components/ui/index.js
@@ -1,0 +1,2 @@
+
+export { default as IconButton } from "./IconButton";
