+++ src/pages/Home/Home.css
@@ -1,0 +1,6 @@
+sections {
+    display: flex; 
+    flex-direction: column; 
+    padding: 1%; 
+    gap: 1%;
+}
